Moved the assertions file into a package prior to breaking apart.
